@@ -451,7 +451,8 @@ describe('entry tests', () => {
     'levels/editors/_all':          './src/sites/studio/pages/levels/editors/_all.js',
     'levels/editors/_dsl':          './src/sites/studio/pages/levels/editors/_dsl.js',
     'projects/index':               './src/sites/studio/pages/projects/index.js',
-    'projects/public':               './src/sites/studio/pages/projects/public.js',
+    'projects/public':            './src/sites/studio/pages/projects/public.js',
+    'projects/featured':          './src/sites/studio/pages/projects/featured.js',
     'schoolInfo':                   './src/sites/studio/pages/schoolInfo.js',
     'schoolInfoInterstitial':       './src/sites/studio/pages/schoolInfoInterstitial.js',
     'scripts/stage_extras':         './src/sites/studio/pages/scripts/stage_extras.js',
