@@ -19,7 +19,7 @@ var CraftVisualizationColumn = function (props) {
         <ArrowButtons />
 
         {props.showFinishButton && <div id="right-button-cell">
-          <button id="rightButton" className="share mc-share-button">
+          <button id="finishButton" className="share mc-share-button">
             <div>{msg.finish()}</div>
           </button>
         </div>}
