@@ -1,13 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import SetupChecklist from '@cdo/apps/lib/kits/maker/ui/SetupChecklist';
-import SetupChecker from '@cdo/apps/lib/kits/maker/util/SetupChecker';
+import SetupGuide from '@cdo/apps/lib/kits/maker/ui/SetupGuide';
 
 $(function () {
-  const setupChecker = new SetupChecker();
   ReactDOM.render(
-      <SetupChecklist setupChecker={setupChecker}/>,
-      document.getElementById('setup-status-mount')
+    <SetupGuide/>,
+    document.getElementById('setup-status-mount')
   );
   $('.maker-setup a').attr('target', '_blank');
 });
