@@ -11,6 +11,7 @@ import ManageStudentsNameCell from './ManageStudentsNameCell';
 import ManageStudentsAgeCell from './ManageStudentsAgeCell';
 import ManageStudentsGenderCell from './ManageStudentsGenderCell';
 import ManageStudentsActionsCell from './ManageStudentsActionsCell';
+import {convertStudentDataToArray} from './manageStudentsRedux';
 import { connect } from 'react-redux';
 
 export const studentSectionDataPropType = PropTypes.shape({
@@ -43,7 +44,7 @@ const nameFormatter = (name, {rowData}) => {
       name={name}
       loginType={rowData.loginType}
       username={rowData.username}
-      isEditing={false}
+      isEditing={rowData.isEditing}
     />
   );
 };
@@ -53,7 +54,7 @@ const ageFormatter = (age, {rowData}) => {
     <ManageStudentsAgeCell
       age={age}
       id={rowData.id}
-      isEditing={false}
+      isEditing={rowData.isEditing}
     />
   );
 };
@@ -63,7 +64,7 @@ const genderFormatter = (gender, {rowData}) => {
     <ManageStudentsGenderCell
       gender={gender}
       id={rowData.id}
-      isEditing={false}
+      isEditing={rowData.isEditing}
     />
   );
 };
@@ -71,20 +72,29 @@ const genderFormatter = (gender, {rowData}) => {
 const passwordFormatter = (loginType, {rowData}) => {
   return (
     <div>
-      {rowData.loginType === SectionLoginType.email &&
-        <PasswordReset
-          resetAction={()=>{}}
-          initialIsResetting={false}
-        />
+      {!rowData.isEditing &&
+        <div>
+          {rowData.loginType === SectionLoginType.email &&
+            <PasswordReset
+              resetAction={()=>{}}
+              initialIsResetting={false}
+            />
+          }
+          {(rowData.loginType === SectionLoginType.word || rowData.loginType === SectionLoginType.picture) &&
+            <ShowSecret
+              resetSecret={()=>{}}
+              initialIsShowing={false}
+              secretWord={rowData.secretWords}
+              secretPicture={rowData.secretPicturePath}
+              loginType={rowData.loginType}
+            />
+          }
+        </div>
       }
-      {(rowData.loginType === SectionLoginType.word || rowData.loginType === SectionLoginType.picture) &&
-        <ShowSecret
-          resetSecret={()=>{}}
-          initialIsShowing={false}
-          secretWord={rowData.secretWords}
-          secretPicture={rowData.secretPicturePath}
-          loginType={rowData.loginType}
-        />
+      {rowData.isEditing &&
+        <div>
+          Auto-generated
+        </div>
       }
     </div>
   );
@@ -94,14 +104,14 @@ const actionsFormatter = function (actions, {rowData}) {
   return (
     <ManageStudentsActionsCell
       id={rowData.id}
-      isEditing={false}
+      isEditing={rowData.isEditing}
     />
   );
 };
 
 class ManageStudentsTable extends Component {
   static propTypes = {
-    //Provided by redux
+    // Provided by redux
     studentData: PropTypes.arrayOf(studentSectionDataPropType),
     loginType: PropTypes.string,
   };
@@ -144,6 +154,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.headerCell,
+            width: 300
           }},
           transforms: [sortable],
         },
@@ -152,6 +163,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.cell,
+            width: 300
           }}
         }
       },
@@ -162,6 +174,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.headerCell,
+            width: 100,
           }},
           transforms: [sortable],
         },
@@ -170,6 +183,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.cell,
+            width: 100,
           }}
         }
       },
@@ -180,6 +194,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.headerCell,
+            width: 150,
           }},
           transforms: [sortable],
         },
@@ -188,6 +203,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.cell,
+            width: 150,
           }}
         }
       },
@@ -201,6 +217,7 @@ class ManageStudentsTable extends Component {
             style: {
             ...tableLayoutStyles.headerCell,
             ...tableLayoutStyles.unsortableHeader,
+            width: 200,
           }},
         },
         cell: {
@@ -208,6 +225,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.cell,
+            width: 200,
           }}
         }
       },
@@ -219,6 +237,7 @@ class ManageStudentsTable extends Component {
             style: {
             ...tableLayoutStyles.headerCell,
             ...tableLayoutStyles.unsortableHeader,
+            width: 200,
           }},
         },
         cell: {
@@ -226,6 +245,7 @@ class ManageStudentsTable extends Component {
           props: {
             style: {
             ...tableLayoutStyles.cell,
+            width: 200,
           }}
         }
       },
@@ -266,5 +286,5 @@ export const UnconnectedManageStudentsTable = ManageStudentsTable;
 
 export default connect(state => ({
   loginType: state.manageStudents.loginType,
-  studentData: state.manageStudents.studentData,
+  studentData: convertStudentDataToArray(state.manageStudents.studentData),
 }))(ManageStudentsTable);
