@@ -1,14 +1,18 @@
 const SET_LOGIN_TYPE = 'manageStudents/SET_LOGIN_TYPE';
 const SET_STUDENTS = 'manageStudents/SET_STUDENTS';
 const SET_SECTION_ID = 'manageStudents/SET_SECTION_ID';
+const START_EDITING_STUDENT = 'manageStudents/START_EDITING_STUDENT';
+const CANCEL_EDITING_STUDENT = 'manageStudents/CANCEL_EDITING_STUDENT';
 
 export const setLoginType = loginType => ({ type: SET_LOGIN_TYPE, loginType });
 export const setSectionId = sectionId => ({ type: SET_SECTION_ID, sectionId});
 export const setStudents = studentData => ({ type: SET_STUDENTS, studentData });
+export const startEditingStudent = (studentId) => ({ type: START_EDITING_STUDENT, studentId });
+export const cancelEditingStudent = (studentId) => ({ type: CANCEL_EDITING_STUDENT, studentId });
 
 const initialState = {
   loginType: '',
-  studentData: [],
+  studentData: {},
   sectionId: null,
 };
 
@@ -31,13 +35,41 @@ export default function manageStudents(state=initialState, action) {
       studentData: action.studentData,
     };
   }
+  if (action.type === START_EDITING_STUDENT) {
+    return {
+      ...state,
+      studentData: {
+        ...state.studentData,
+        [action.studentId]: {
+          ...state.studentData[action.studentId],
+          isEditing: true
+        }
+      }
+    };
+  }
+  if (action.type === CANCEL_EDITING_STUDENT) {
+    return {
+      ...state,
+      studentData: {
+        ...state.studentData,
+        [action.studentId]: {
+          ...state.studentData[action.studentId],
+          isEditing: false
+        }
+      }
+    };
+  }
 
   return state;
 }
 
+// Converts data from /v2/sections/sectionid/students to a set of key/value
+// objects for the redux store
 export const convertStudentServerData = (studentData, loginType, sectionId) => {
-  return studentData.map((student) => {
-    return {
+  let studentLookup = {};
+  for (let i=0; i < studentData.length; i++) {
+    let student = studentData[i];
+    studentLookup[student.id] = {
       id: student.id,
       name: student.name,
       username: student.username,
@@ -47,6 +79,15 @@ export const convertStudentServerData = (studentData, loginType, sectionId) => {
       secretPicturePath: student.secret_picture_path,
       loginType: loginType,
       sectionId: sectionId,
+      isEditing: false,
     };
-  });
+  }
+  return studentLookup;
+};
+
+// Converts key/value id/student pairs to an array of student objects for the
+// component to display
+// TODO(caleybrock): memoize this - sections could be a few thousand students
+export const convertStudentDataToArray = (studentData) => {
+  return Object.values(studentData);
 };
