@@ -250,19 +250,11 @@ class ApiController < ApplicationController
     section = load_section
     script = load_script(section)
 
-    data = {}
-    # TODO: This could likely be constructed more efficiently. At the very least,
-    # instead of asking for a summary, and then using only one portion of it (levels)
-    # we could probably expose a way to get just levels and have it be in the same
-    # form as user_progress. However, we might be able to do even better and query
-    # all the data that we need in a single db request
     # TODO: We'll want to support some form of pagination for this API. One option
     # would be to imitate the approach used by the section_progress API, however
     # that has some limitations and was largely meant as a quick and dirty fix for
     # pagination when it was implemented
-    section.students.each do |student|
-      data[student.id] = summarize_user_progress(script, student)[:levels]
-    end
+    data = get_level_progress(section.students, script)
     render json: data
   end
 
@@ -292,23 +284,6 @@ class ApiController < ApplicationController
   def script_structure
     script = Script.get_from_cache(params[:script])
     render json: script.summarize
-  end
-
-  # Return a JSON summary of the user's progress across all scripts.
-  def user_progress_for_all_scripts
-    # I do not believe this API is used anywhere. I would like to remove it, but
-    # as a first step, I thought I would add some logging
-    Honeybadger.notify(
-      error_message: "user_progress_for_all_scripts called",
-      error_class: "ApiController.user_progress_for_all_scripts",
-    )
-
-    user = current_user
-    if user
-      render json: summarize_user_progress_for_all_scripts(user)
-    else
-      render json: {}
-    end
   end
 
   # Return a JSON summary of the user's progress for params[:script].
