@@ -16,7 +16,7 @@
 #  address_line3      :string(30)
 #  latitude           :decimal(8, 6)
 #  longitude          :decimal(9, 6)
-#  state_school_id    :string(11)
+#  state_school_id    :string(255)
 #
 # Indexes
 #
@@ -36,7 +36,7 @@ class School < ActiveRecord::Base
 
   has_many :school_stats_by_year
 
-  validates :state_school_id, allow_blank: true, length: {is: 11}, format: {with: /\A[A-Z]{2}-[0-9]{3}-[0-9]{4}\z/, message: "must be {State Code}-xxx-xxxx where each x is a numeral"}
+  validates :state_school_id, allow_blank: true, format: {with: /\A[A-Z]{2}-.+-.+\z/, message: "must be {State Code}-{State District Id}-{State School Id}"}
 
   # Gets the full address of the school.
   # @return [String] The full address.
