@@ -31,8 +31,26 @@ module Pd::Application
     end
 
     def find_teachercon_workshop(course:, city:, year: 2018)
+      find_scheduled_workshop(
+        course: course,
+        subject: Pd::Workshop::SUBJECT_TEACHER_CON,
+        city: city,
+        year: year
+      )
+    end
+
+    def find_fit_workshop(course:, city:, year: 2018)
+      find_scheduled_workshop(
+        course: course,
+        subject: Pd::Workshop::SUBJECT_FIT,
+        city: city,
+        year: year
+      )
+    end
+
+    def find_scheduled_workshop(course:, subject:, city:, year: 2018)
       Pd::Workshop.
-        where(course: course, subject: Pd::Workshop::SUBJECT_TEACHER_CON).
+        where(course: course, subject: subject).
         scheduled_start_on_or_after(Date.new(year)).
         scheduled_start_on_or_before(Date.new(year + 1)).
         where('location_address like ?', "%#{city}%").
