@@ -614,7 +614,7 @@ FactoryGirl.define do
   end
 
   factory :pd_facilitator1819_application, class: 'Pd::Application::Facilitator1819Application' do
-    association :user, factory: :teacher, strategy: :create
+    association :user, factory: [:teacher, :with_school_info], strategy: :create
     course 'csp'
     transient do
       form_data_hash {build :pd_facilitator1819_application_hash, program: Pd::Application::Facilitator1819Application::PROGRAMS[course.to_sym]}
@@ -816,6 +816,10 @@ FactoryGirl.define do
               liveFarAway: "Yes",
               needHotel: "No",
               photoRelease: "Yes",
+              howOfferCsp: "As an AP course",
+              haveTaughtAp: "Yes",
+              haveTaughtWrittenProjectCourse: "Yes",
+              gradingSystem: 'Numerical and/or letter grades (e.g., 0 - 100% or F- A)',
             }
           )
         end
