@@ -2,6 +2,7 @@ require 'fileutils'
 require 'net/http'
 require 'net/http/responses'
 require 'uri'
+require 'cdo/circle_utils'
 require 'cdo/slack'
 
 # This class is intended to be a thin wrapper around our chat client
@@ -11,7 +12,11 @@ class ChatClient
   @@logger = nil
 
   def self.log(message, options={})
-    message(CDO.slack_log_room, message, options)
+    if CircleUtils.circle?
+      CDO.log.info("[#{CDO.slack_log_room}] #{message}")
+    else
+      message(CDO.slack_log_room, message, options)
+    end
   end
 
   # @param room [String] The room to post to which message should be posted.
